refactor: consolidate preprocessing and move workflow helpers to workflows layer
