+++ simba/workflows/training.py
@@ -19,7 +19,6 @@ import simba.core.data.spectrum
 from simba.core.chemistry.mces_loader.load_mces import LoadMCES
 from simba.core.data.molecule_pairs import MoleculePairsOpt
 from simba.core.data.multitask_dataset_builder import MultitaskDataLoader
-from simba.core.data.preprocessing_simba import PreprocessingSimba
 from simba.core.data.sampling.custom_weighted_random_sampler import (
     CustomWeightedRandomSampler,
 )
@@ -29,6 +28,7 @@ from simba.core.training.losscallback import LossCallback
 from simba.core.training.train_utils import TrainUtils
 from simba.utils.logger_setup import logger
 from simba.utils.sanity_checks import SanityChecks
+from simba.workflows.utils import load_spectra
 
 
 # Backward compatibility: Support loading old pickle files with old module paths
@@ -84,7 +84,7 @@ def load_dataset(cfg: DictConfig):
         )
 
         mgf_path = mapping["mgf_path"]
-        all_spectra = PreprocessingSimba.load_spectra(mgf_path, cfg)
+        all_spectra = load_spectra(mgf_path, cfg)
 
         # Create spectrum lookup by MGF index
         spectra_by_idx = {s.mgf_index: s for s in all_spectra}
